Remove Notification Goal Event On Item You Already have On Inventory Sync And Save (KnownTech Entries,PDAScanner Entries,PDAEncyclopediaEntries ) All Progress of Encyclopedia and BluePrint Fixed Add Comand (mpsave) to force save game on server Add Notifiacion Sync of Partial Scan Of Item PowerLevelChangedProcessor Bug Drain Loop Fix
