+++ NitroxClient/GameLogic/Item.cs
@@ -47,18 +47,9 @@ namespace NitroxClient.GameLogic
 
             Log.Debug("Dropping item with guid: " + guid);
 
-            if (techType == TechType.Beacon | techType == TechType.Pipe)
-            {
-                DroppedItemPacket droppedItem = new DroppedItemPacket(guid, waterpark, techType, dropPosition, bytes);
-                packetSender.Send(droppedItem);
-            }
-            else
-            {
-                DroppedItem droppedItem = new DroppedItem(guid, waterpark, techType, dropPosition, bytes);
-                packetSender.Send(droppedItem);
-            }
+            DroppedItem droppedItem = new DroppedItem(guid, waterpark, techType, dropPosition, bytes);
+            packetSender.Send(droppedItem);
 
-            
         }
 
         private Optional<string> GetCurrentWaterParkGuid()
